time_t is 64 bits on 64-bit architectures on Linux

+++ include/kernel/linux/sys/resource.h
@@ -1,5 +1,5 @@
 /* $Id$ */
-/* Copyright (c) 2009 Pierre Pronchery <khorben@defora.org> */
+/* Copyright (c) 2007-2013 Pierre Pronchery <khorben@defora.org> */
 /* This file is part of DeforaOS System libc */
 /* This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -31,10 +31,10 @@ typedef int suseconds_t;
 # endif
 # ifndef time_t
 #  define time_t time_t
-typedef int time_t;
+typedef long time_t;
 # endif
 # ifndef timeval
-# define timeval timeval
+#  define timeval timeval
 struct timeval
 {
 	time_t tv_sec;
