+++ include/kernel/linux/time.h
@@ -1,5 +1,5 @@
 /* $Id$ */
-/* Copyright (c) 2007-2012 Pierre Pronchery <khorben@defora.org> */
+/* Copyright (c) 2007-2013 Pierre Pronchery <khorben@defora.org> */
 /* This file is part of DeforaOS System libc */
 /* This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -30,7 +30,7 @@ typedef int clockid_t;
 # endif
 # ifndef time_t
 #  define time_t time_t
-typedef int time_t;
+typedef long time_t;
 # endif
 
 #endif /* !LIBC_KERNEL_LINUX_TIME_H */
